+++ src/rules/h002_shortest_subject_line.rs
@@ -1,15 +1,29 @@
 use crate::achievement::{Achievement, Rule, RuleFactory};
+use crate::config::RulesConfig;
 
 /// The shortest subject line in a branch
+#[derive(Default)]
 pub struct ShortestSubjectLine {
-    length_threshold: usize,
+    config: H002Config,
     shortest_so_far: Option<(git2::Oid, usize)>,
 }
 
-fn shortest_subject_line() -> Box<dyn Rule> {
+#[derive(Debug, Clone, PartialEq, serde::Deserialize, serde::Serialize)]
+pub struct H002Config {
+    pub length_threshold: usize,
+}
+
+impl Default for H002Config {
+    fn default() -> Self {
+        Self {
+            length_threshold: 10,
+        }
+    }
+}
+
+fn shortest_subject_line(config: &RulesConfig) -> Box<dyn Rule> {
     Box::new(ShortestSubjectLine {
-        // TODO: Make this configurable (#58)
-        length_threshold: 10,
+        config: config.h2_shortest_subject_line.clone().unwrap_or_default(),
         shortest_so_far: None,
     })
 }
@@ -37,7 +51,7 @@ impl Rule for ShortestSubjectLine {
     }
     fn process(&mut self, commit: &git2::Commit, _repo: &git2::Repository) -> Option<Achievement> {
         let length = subject_length(commit);
-        if length < self.length_threshold {
+        if length < self.config.length_threshold {
             match self.shortest_so_far {
                 Some((_, shortest_length)) => {
                     if length < shortest_length {
@@ -70,8 +84,14 @@ mod tests {
 
     #[test]
     fn test_all_above_threshold() {
+        let config = RulesConfig {
+            h2_shortest_subject_line: Some(H002Config {
+                length_threshold: 7,
+            }),
+            ..Default::default()
+        };
         let repo = fixtures::repository::with_empty_commits(&["0123456789", "1234567890"]).unwrap();
-        let rules = vec![shortest_subject_line()];
+        let rules = vec![shortest_subject_line(&config)];
         let achievements = grant_with_rules("HEAD", &repo.repo, rules).unwrap();
         let achievements: Vec<_> = achievements.collect();
         assert!(achievements.is_empty());
@@ -81,7 +101,7 @@ mod tests {
     fn test_has_short_subject() {
         let repo =
             fixtures::repository::with_empty_commits(&["0123456789", "1234567", "1234"]).unwrap();
-        let rules = vec![shortest_subject_line()];
+        let rules = vec![Box::new(ShortestSubjectLine::default()) as Box<dyn Rule>];
         let achievements = grant_with_rules("HEAD", &repo.repo, rules).unwrap();
         let achievements: Vec<_> = achievements.collect();
         assert_eq!(achievements.len(), 1);
@@ -98,10 +118,10 @@ mod tests {
         let repo2 =
             fixtures::repository::with_empty_commits(&["1234567890", "2345671", "1234"]).unwrap();
 
-        let rules1 = vec![shortest_subject_line()];
+        let rules1 = vec![Box::new(ShortestSubjectLine::default()) as Box<dyn Rule>];
         // grant_with_rules() consumes the rules Vec, so there _can't_ be any state held between
         // processing any two repositories
-        let rules2 = vec![shortest_subject_line()];
+        let rules2 = vec![Box::new(ShortestSubjectLine::default()) as Box<dyn Rule>];
 
         let achievements1 = grant_with_rules("HEAD", &repo1.repo, rules1).unwrap();
         let achievements2 = grant_with_rules("HEAD", &repo2.repo, rules2).unwrap();